+++ Template_Project/Src/datatypes.h
@@ -2,9 +2,9 @@
 #define DATATYPES_H
 
 typedef struct coordinates {
-	int x;
-	int y;
-	short player;
+	WORD x;
+	WORD y;
+	BYTE player;
 } Coordinates;
 
 #endif
